# docs/executor-api.openapi.yaml
openapi: 3.1.0
info:
  title: Dillflix Playback Executor API
  version: 1.0.0
  description: 'Implemented in dillflix-live-controller 0.9.0. Three operations: asynchronous Play, token status
    including lifecycle, and Cancel including active stop and device intent fencing. Configure PLAYBACK_ADAPTER=prime-video
    and EXECUTOR_API_TOKEN. See executor-api-handoff.md and executor-setup.md.'
servers:
- url: /
  description: Same origin as the configured controller deployment
security:
- ServiceBearer: []
paths:
  /v1/playbacks:
    post:
      operationId: initiatePlayback
      summary: Durably accept live playback and return its token
      description: 'Identical request_id and body retries return the same token without relaunching. Changed body
        is 409. A transport timeout is an unknown outcome: retry this same POST while current intent remains eligible.
        No separate request-ID recovery endpoint. Higher intents stop older input before navigation starts. Check
        the durable device cancellation fence before issuing any input.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackRequest'
            examples:
              StartPlayback:
                $ref: '#/components/examples/StartPlayback'
      responses:
        '202':
          description: Accepted; navigation runs asynchronously. Acceptance does not verify playback.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackReport'
              examples:
                Accepted:
                  $ref: '#/components/examples/Accepted'
          headers:
            Location:
              schema:
                type: string
              description: Relative token status URL, e.g. /v1/playbacks/pb-example-001.
        '200':
          description: Identical retry; same token and current report, no duplicate navigation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackReport'
              examples:
                Verified:
                  $ref: '#/components/examples/Verified'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '413':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/playbacks/{token}:
    get:
      operationId: getPlaybackStatus
      summary: Read playback progress, current evidence and event completion by token
      description: 'Read-only, bounded and Cache-Control: no-store. Never navigate, resume, or stop from this getter.
        A successful status read can report unavailable playback/lifecycle evidence. A historical playing_verified
        result is not proof of current playback. Ended requires explicit evidence for the requested content scope.'
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
      responses:
        '200':
          description: Complete status envelope; independent evidence may be unknown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackReport'
              examples:
                Verified:
                  $ref: '#/components/examples/Verified'
                Ended:
                  $ref: '#/components/examples/Ended'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/playbacks/cancel:
    post:
      operationId: cancelPlayback
      summary: Cancel navigation and stop owned active playback, or fence a device for manual control
      description: 'One Cancel API, two mutually exclusive scopes. Persist cancellation before stopping/draining
        input. Return 200 only after confirmed quiescence and active stop/already-inactive. Retry is idempotent.
        On timeout/503 cancellation is unconfirmed: manual input stays disabled and the controller retries. A device
        fence covers late Play deliveries whose token was never received. No authority or renewal endpoints.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
            examples:
              CancelToken:
                $ref: '#/components/examples/CancelToken'
              CancelForManual:
                $ref: '#/components/examples/CancelForManual'
      responses:
        '200':
          description: Durable cancellation acknowledged; no further input in scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResult'
              examples:
                CancelledToken:
                  $ref: '#/components/examples/CancelledToken'
                ManualReady:
                  $ref: '#/components/examples/ManualReady'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
        '413':
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    ServiceBearer:
      type: http
      scheme: bearer
      description: Server-side scoped service credential, separate from playback tokens and browser owner_token.
  schemas:
    CancelRequest:
      properties:
        device_id:
          type: string
          const: living-room
          title: Device Id
        token:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Token
        through_intent_version:
          anyOf:
          - type: integer
            maximum: 9007199254740991.0
            minimum: 0.0
          - type: 'null'
          title: Through Intent Version
      additionalProperties: false
      type: object
      required:
      - device_id
      title: CancelRequest
    CancelResult:
      properties:
        device_id:
          type: string
          const: living-room
          title: Device Id
        token:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
        through_intent_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Through Intent Version
        input_quiescent:
          type: boolean
          const: true
          title: Input Quiescent
        active_playback:
          type: string
          enum:
          - stopped
          - already_inactive
          title: Active Playback
        acknowledged_at:
          type: string
          format: date-time
          title: Acknowledged At
      additionalProperties: false
      type: object
      required:
      - device_id
      - token
      - through_intent_version
      - input_quiescent
      - active_playback
      - acknowledged_at
      title: CancelResult
    Cancellation:
      properties:
        state:
          type: string
          enum:
          - none
          - requested
          - acknowledged
          title: State
        input_quiescent:
          type: boolean
          title: Input Quiescent
      additionalProperties: false
      type: object
      required:
      - state
      - input_quiescent
      title: Cancellation
    ContentStatus:
      properties:
        content_id:
          type: string
          title: Content Id
        lookup_state:
          type: string
          enum:
          - ok
          - unavailable
          title: Lookup State
        effective_state:
          type: string
          enum:
          - scheduled
          - live
          - ended
          - cancelled
          - delayed
          - suspended
          - postponed
          - unknown
          title: Effective State
        stale:
          type: boolean
          title: Stale
        observation:
          anyOf:
          - $ref: '#/components/schemas/LifecycleObservation'
          - type: 'null'
        checked_at:
          type: string
          format: date-time
          title: Checked At
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - content_id
      - lookup_state
      - effective_state
      - stale
      - observation
      - checked_at
      - error
      title: ContentStatus
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        retryable:
          type: boolean
          title: Retryable
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - retryable
      title: ErrorDetail
    Evidence:
      properties:
        method:
          type: string
          const: device_observation
          title: Method
        evidence_id:
          type: string
          title: Evidence Id
        summary:
          type: string
          title: Summary
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        captured_at:
          type: string
          format: date-time
          title: Captured At
        decision:
          anyOf:
          - type: string
            const: confirmed
          - type: 'null'
          title: Decision
      additionalProperties: false
      type: object
      required:
      - method
      - evidence_id
      - summary
      - confidence
      - captured_at
      title: Evidence
    LifecycleObservation:
      properties:
        content_id:
          type: string
          title: Content Id
        state:
          type: string
          enum:
          - scheduled
          - live
          - ended
          - cancelled
          - delayed
          - suspended
          - postponed
          - unknown
          title: State
        source:
          type: string
          enum:
          - teamarr_feed
          - prime_video_visual
          title: Source
        simulated:
          type: boolean
          const: false
          title: Simulated
        timestamp_basis:
          type: string
          enum:
          - feed_received
          - device_observed
          title: Timestamp Basis
        observed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Observed At
        received_at:
          type: string
          format: date-time
          title: Received At
        valid_until:
          type: string
          format: date-time
          title: Valid Until
        evidence:
          anyOf:
          - $ref: '#/components/schemas/Evidence'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - content_id
      - state
      - source
      - simulated
      - timestamp_basis
      - observed_at
      - received_at
      - valid_until
      title: LifecycleObservation
    ObservationStatus:
      properties:
        state:
          type: string
          enum:
          - fresh
          - stale
          - unavailable
          title: State
        checked_at:
          type: string
          format: date-time
          title: Checked At
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - state
      - checked_at
      - error
      title: ObservationStatus
    Operation:
      properties:
        state:
          type: string
          enum:
          - accepted
          - navigating
          - playing_verified
          - failed
          - cancelled
          - superseded
          - timed_out
          title: State
        phase:
          type: string
          title: Phase
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        deadline_at:
          type: string
          format: date-time
          title: Deadline At
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
        attempts:
          items:
            $ref: '#/components/schemas/RouteAttempt'
          type: array
          title: Attempts
      additionalProperties: false
      type: object
      required:
      - state
      - phase
      - created_at
      - updated_at
      - finished_at
      - deadline_at
      - error
      - attempts
      title: Operation
    PlaybackObservation:
      properties:
        device_id:
          type: string
          const: living-room
          title: Device Id
        request_id:
          type: string
          title: Request Id
        intent_version:
          type: integer
          title: Intent Version
        content_id:
          type: string
          title: Content Id
        viewing_option_id:
          type: string
          title: Viewing Option Id
        presentation:
          type: string
          const: live
          title: Presentation
        verified:
          type: boolean
          title: Verified
        simulated:
          type: boolean
          const: false
          title: Simulated
        health:
          type: string
          const: healthy
          title: Health
        observed_at:
          type: string
          format: date-time
          title: Observed At
        valid_until:
          type: string
          format: date-time
          title: Valid Until
        evidence:
          $ref: '#/components/schemas/Evidence'
      additionalProperties: false
      type: object
      required:
      - device_id
      - request_id
      - intent_version
      - content_id
      - viewing_option_id
      - presentation
      - verified
      - simulated
      - health
      - observed_at
      - valid_until
      - evidence
      title: PlaybackObservation
    PlaybackReport:
      properties:
        schema_version:
          type: integer
          const: 1
          title: Schema Version
        token:
          type: string
          title: Token
        request_id:
          type: string
          title: Request Id
        device_id:
          type: string
          const: living-room
          title: Device Id
        content_id:
          type: string
          title: Content Id
        intent_version:
          type: integer
          title: Intent Version
        revision:
          type: integer
          title: Revision
        operation:
          $ref: '#/components/schemas/Operation'
        observation:
          anyOf:
          - $ref: '#/components/schemas/PlaybackObservation'
          - type: 'null'
        observation_status:
          $ref: '#/components/schemas/ObservationStatus'
        content_status:
          $ref: '#/components/schemas/ContentStatus'
        cancellation:
          $ref: '#/components/schemas/Cancellation'
        retained_until:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Retained Until
      additionalProperties: false
      type: object
      required:
      - schema_version
      - token
      - request_id
      - device_id
      - content_id
      - intent_version
      - revision
      - operation
      - observation
      - observation_status
      - content_status
      - cancellation
      - retained_until
      title: PlaybackReport
    PlaybackRequest:
      properties:
        schema_version:
          type: integer
          const: 1
          title: Schema Version
          default: 1
        request_id:
          type: string
          maxLength: 200
          minLength: 1
          title: Request Id
        device_id:
          type: string
          const: living-room
          title: Device Id
        intent_version:
          type: integer
          maximum: 9007199254740991.0
          minimum: 0.0
          title: Intent Version
        content_id:
          type: string
          maxLength: 1024
          minLength: 1
          title: Content Id
        mode:
          type: string
          const: live
          title: Mode
        purpose:
          type: string
          enum:
          - selection
          - route_handoff
          - recovery
          title: Purpose
          default: selection
        previous_request_id:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Previous Request Id
        content_snapshot_schema_version:
          type: integer
          const: 1
          title: Content Snapshot Schema Version
          default: 1
        content_snapshot:
          additionalProperties: true
          type: object
          title: Content Snapshot
        allowed_viewing_options:
          items:
            additionalProperties: true
            type: object
          type: array
          maxItems: 100
          minItems: 1
          title: Allowed Viewing Options
        deadline_at:
          type: string
          format: date-time
          title: Deadline At
      additionalProperties: false
      type: object
      required:
      - request_id
      - device_id
      - intent_version
      - content_id
      - mode
      - content_snapshot
      - allowed_viewing_options
      - deadline_at
      title: PlaybackRequest
    RouteAttempt:
      properties:
        attempt:
          type: integer
          title: Attempt
        viewing_option_id:
          type: string
          title: Viewing Option Id
        phase:
          type: string
          title: Phase
        started_at:
          type: string
          format: date-time
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - attempt
      - viewing_option_id
      - phase
      - started_at
      - finished_at
      - error
      title: RouteAttempt
    Problem:
      additionalProperties: false
      properties:
        type:
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          title: Detail
          type: string
        code:
          title: Code
          type: string
        retryable:
          title: Retryable
          type: boolean
      required:
      - type
      - title
      - status
      - detail
      - code
      - retryable
      title: Problem
      type: object
  examples:
    StartPlayback:
      summary: Synthetic full snapshot; substitute the untouched real Teamarr entry and options
      value:
        schema_version: 1
        request_id: req-example-001
        device_id: living-room
        intent_version: 23
        content_id: example-content-001
        mode: live
        purpose: selection
        previous_request_id: null
        content_snapshot_schema_version: 1
        content_snapshot:
          id: example-content-001
          kind: event
          source: games
          title: Canadiens at Maple Leafs
          competition: nhl
          sports:
          - nhl
          provider: example
          start_time: '2026-09-30T19:00:00Z'
          expected_end_time: '2026-09-30T22:00:00Z'
          end_time_estimated: true
          timing_basis: demo_duration
          status: live
          event:
            league: nhl
            season_type: playoffs
            away_team_details:
              id: MTL
              provider: demo
              city: Montreal
              name: Canadiens
              full_name: Montreal Canadiens
              short_name: Canadiens
              abbreviation: MTL
              logo_url: null
            home_team_details:
              id: TOR
              provider: demo
              city: Toronto
              name: Maple Leafs
              full_name: Toronto Maple Leafs
              short_name: Maple Leafs
              abbreviation: TOR
              logo_url: null
            event_id: canadiens
            provider: demo
            home_team: Toronto Maple Leafs
            away_team: Montreal Canadiens
          sessions: []
          broadcast: null
          related_ids: []
          viewing_options:
          - id: example-option-001
            app: demo_provider
            decision: eligible
            reasons: []
            presentation: live
            basis: synthetic_contract_example
            coverage_type: full
          artwork: {}
        allowed_viewing_options:
        - id: example-option-001
          app: demo_provider
          decision: eligible
          reasons: []
          presentation: live
          basis: synthetic_contract_example
          coverage_type: full
        deadline_at: '2026-09-30T20:02:00Z'
      x-schema: PlaybackRequest
    Accepted:
      summary: Durably accepted; no playback claim
      value:
        schema_version: 1
        token: pb-example-001
        request_id: req-example-001
        device_id: living-room
        content_id: example-content-001
        intent_version: 23
        revision: 1
        operation:
          state: accepted
          phase: queued
          created_at: '2026-09-30T20:00:00Z'
          updated_at: '2026-09-30T20:00:00Z'
          finished_at: null
          deadline_at: '2026-09-30T20:02:00Z'
          error: null
          attempts: []
        observation: null
        observation_status:
          state: unavailable
          checked_at: '2026-09-30T20:00:10Z'
          error: null
        content_status:
          content_id: example-content-001
          lookup_state: unavailable
          effective_state: unknown
          stale: true
          observation: null
          checked_at: '2026-09-30T20:00:10Z'
          error:
            code: status_source_unavailable
            message: Lifecycle evidence is unavailable
            retryable: true
        cancellation:
          state: none
          input_quiescent: false
        retained_until: null
      x-schema: PlaybackReport
    Verified:
      summary: Fresh playback independently verified while lifecycle is unavailable
      value:
        schema_version: 1
        token: pb-example-001
        request_id: req-example-001
        device_id: living-room
        content_id: example-content-001
        intent_version: 23
        revision: 7
        operation:
          state: playing_verified
          phase: verified
          created_at: '2026-09-30T20:00:00Z'
          updated_at: '2026-09-30T20:00:09Z'
          finished_at: '2026-09-30T20:00:09Z'
          deadline_at: '2026-09-30T20:02:00Z'
          error: null
          attempts: []
        observation:
          device_id: living-room
          request_id: req-example-001
          intent_version: 23
          content_id: example-content-001
          viewing_option_id: example-option-001
          presentation: live
          verified: true
          simulated: false
          health: healthy
          observed_at: '2026-09-30T20:00:10Z'
          valid_until: '2026-09-30T20:00:25Z'
          evidence:
            method: device_observation
            evidence_id: obs-example-042
            summary: Synthetic example of matched content identity and live presentation
            confidence: null
            captured_at: '2026-09-30T20:00:10Z'
        observation_status:
          state: fresh
          checked_at: '2026-09-30T20:00:10Z'
          error: null
        content_status:
          content_id: example-content-001
          lookup_state: unavailable
          effective_state: unknown
          stale: true
          observation: null
          checked_at: '2026-09-30T20:00:10Z'
          error:
            code: status_source_unavailable
            message: Lifecycle evidence is unavailable
            retryable: true
        cancellation:
          state: none
          input_quiescent: false
        retained_until: null
      x-schema: PlaybackReport
    Ended:
      summary: Confirmed game completion despite stale playback evidence
      value:
        schema_version: 1
        token: pb-example-001
        request_id: req-example-001
        device_id: living-room
        content_id: example-content-001
        intent_version: 23
        revision: 20
        operation:
          state: playing_verified
          phase: verified
          created_at: '2026-09-30T20:00:00Z'
          updated_at: '2026-09-30T20:00:09Z'
          finished_at: '2026-09-30T20:00:09Z'
          deadline_at: '2026-09-30T20:02:00Z'
          error: null
          attempts: []
        observation:
          device_id: living-room
          request_id: req-example-001
          intent_version: 23
          content_id: example-content-001
          viewing_option_id: example-option-001
          presentation: live
          verified: false
          simulated: false
          health: healthy
          observed_at: '2026-09-30T20:00:10Z'
          valid_until: '2026-09-30T20:00:25Z'
          evidence:
            method: device_observation
            evidence_id: obs-example-042
            summary: Synthetic example of matched content identity and live presentation
            confidence: null
            captured_at: '2026-09-30T20:00:10Z'
        observation_status:
          state: stale
          checked_at: '2026-09-30T22:05:01Z'
          error: null
        content_status:
          content_id: example-content-001
          lookup_state: ok
          effective_state: ended
          stale: false
          observation:
            content_id: example-content-001
            state: ended
            source: prime_video_visual
            simulated: false
            timestamp_basis: device_observed
            observed_at: '2026-09-30T22:05:00Z'
            received_at: '2026-09-30T22:05:01Z'
            valid_until: '2026-09-30T22:07:00Z'
            evidence:
              method: device_observation
              evidence_id: final-example-001
              summary: 'Synthetic example: two readings at least 15 seconds apart show FINAL for this exact game'
              confidence: null
              decision: confirmed
              captured_at: '2026-09-30T22:05:00Z'
          checked_at: '2026-09-30T22:05:01Z'
          error: null
        cancellation:
          state: none
          input_quiescent: false
        retained_until: null
      x-schema: PlaybackReport
    CancelToken:
      value:
        device_id: living-room
        token: pb-example-001
      x-schema: CancelRequest
    CancelForManual:
      value:
        device_id: living-room
        through_intent_version: 24
      x-schema: CancelRequest
    CancelledToken:
      value:
        device_id: living-room
        token: pb-example-001
        through_intent_version: null
        input_quiescent: true
        active_playback: stopped
        acknowledged_at: '2026-09-30T20:05:00Z'
      x-schema: CancelResult
    ManualReady:
      value:
        device_id: living-room
        token: null
        through_intent_version: 24
        input_quiescent: true
        active_playback: stopped
        acknowledged_at: '2026-09-30T20:05:00Z'
      x-schema: CancelResult
    Conflict:
      summary: No input allowed after ownership change
      value:
        type: urn:dillflix:problem:stale_intent
        title: Playback intent conflict
        status: 409
        detail: The device intent is at or below its durable cancellation fence.
        code: stale_intent
        retryable: false
      x-schema: Problem
  responses:
    Problem:
      description: Structured failure. No new navigation effects for rejected mutations.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            Conflict:
              $ref: '#/components/examples/Conflict'